+++ src/acao/Acao.java
@@ -1,0 +1,10 @@
+package acao;
+
+public class Acao {
+  public final static Bloqueio BLOCK = new Bloqueio();
+  public final static Inverter INVERTER = new Inverter();
+  public final static Mais2 MAIS2 = new Mais2();
+  public final static Mais4 MAIS4 = new Mais4();
+  public final static TrocaCor TROCACOR = new TrocaCor();
+
+}
