+++ src/acao/Bloqueio.java
@@ -1,0 +1,10 @@
+package acao;
+
+public class Bloqueio extends Acao {
+  public Bloqueio() {
+  }
+
+  public int block() {
+    return 1;
+  }
+}
