+++ src/acao/Inverter.java
@@ -1,0 +1,10 @@
+package acao;
+
+public class Inverter extends Acao {
+  public Inverter() {
+  }
+
+  public int inversao() {
+    return -1;
+  }
+}
