+++ src/acao/Mais2.java
@@ -1,0 +1,11 @@
+package acao;
+
+public class Mais2 extends Acao {
+
+  public Mais2() {
+  }
+
+  public int Soma() {
+    return 2;
+  }
+}
