+++ src/acao/TrocaCor.java
@@ -1,0 +1,17 @@
+package acao;
+
+//import java.util.Scanner;
+
+public class TrocaCor extends Acao {
+  public TrocaCor() {
+  }
+
+  public String mudarCor(String cor) {
+    /*
+     * Scanner sc;
+     * sc = new Scanner(System.in);
+     * cor = sc.nextLine();
+     */
+    return cor;
+  }
+}
