--- src/base/Acao/Acao.java
@@ -1,10 +1,0 @@
-package Acao;
-
-public class Acao {
-  public final static Bloqueio BLOCK = new Bloqueio();
-  public final static Inverter INVERTER = new Inverter();
-  public final static Mais2 MAIS2 = new Mais2();
-  public final static Mais4 MAIS4 = new Mais4();
-  public final static TrocaCor TROCACOR = new TrocaCor();
-
-}
