--- src/base/Acao/Bloqueio.java
@@ -1,10 +1,0 @@
-package Acao;
-
-public class Bloqueio extends Acao {
-  public Bloqueio() {
-  }
-
-  public int block() {
-    return 1;
-  }
-}
