--- src/base/Acao/Inverter.java
@@ -1,10 +1,0 @@
-package Acao;
-
-public class Inverter extends Acao {
-  public Inverter() {
-  }
-
-  public int inversao() {
-    return -1;
-  }
-}
