--- src/base/Acao/Mais2.java
@@ -1,11 +1,0 @@
-package Acao;
-
-public class Mais2 extends Acao {
-
-  public Mais2() {
-  }
-
-  public int Soma() {
-    return 2;
-  }
-}
