--- src/base/Acao/Mais4.java
@@ -1,20 +1,0 @@
-package Acao;
-//import java.util.Scanner;
-
-public class Mais4 extends Acao {
-  public Mais4() {
-  }
-
-  public String mudarCor(String cor) {
-    /*
-     * Scanner sc;
-     * sc = new Scanner(System.in);
-     * cor = sc.nextLine();
-     */
-    return cor;
-  }
-
-  public int Soma() {
-    return 4;
-  }
-}
