--- src/base/Acao/TrocaCor.java
@@ -1,17 +1,0 @@
-package Acao;
-
-//import java.util.Scanner;
-
-public class TrocaCor extends Acao {
-  public TrocaCor() {
-  }
-
-  public String mudarCor(String cor) {
-    /*
-     * Scanner sc;
-     * sc = new Scanner(System.in);
-     * cor = sc.nextLine();
-     */
-    return cor;
-  }
-}
