+++ src/base/MaoCartas.java
@@ -1,15 +1,13 @@
-package jogador;
-
-import baralho.Carta;
+package base;
 
 public class MaoCartas {
-    Carta[] cartas = new Carta[112];
-    
-    public void contaCartas() {
-        
-    }
+	Carta[] cartas = new Carta[112];
 
-    public Carta[] getCartas() {
-        return cartas;
-    }
+	public void contaCartas() {
+
+	}
+
+	public Carta[] getCartas() {
+		return cartas;
+	}
 }
