--- src/base/baralho.java
@@ -1,28 +1,0 @@
-//criando um baralho
-import java.util.Random;
-
-public class Baralho{
-  private Carta[] baralho;
-  private pilhaDescarte[] baralho;
-  //atributos: quantidade de cartas
-  //métodos: comprar cartas, verificar quantidade de cartas
-
-  private pilhaCompra[] baralho;
-  //atributos: última carta
-  //métodos: embaralhar
-  
-  baralho = new Carta[112];
-  
-
-  public void inicializarBaralho(){
-  }
-  public void embaralhar(){
-  }
-  public void distribuir(){
-  }
-  public void comprarCarta(){
-  }
-  public void quantCarta(){   //para verificar a quantidade de cartas
-  }
-  
-}
