--- src/base/cartas/CartaComCor.java
@@ -1,12 +1,0 @@
-package cartas;
-
-public class CartaComCor extends Carta {
-
-	
-	
-	public String getAcao() {
-		return "Esta carta não possui ação";
-	}
-	
-}
-
