--- src/base/cartas/CartaNormal.java
@@ -1,17 +1,0 @@
-package cartas;
-
-public class CartaNormal extends CartaComCor {
-	
-	public CartaNormal(String c, int n){//Carta normal com cor
-		super.setCor(c);
-		super.setNumero(n);
-		super.setAcao(null);
-	}
-	
-	
-	
-	
-	
-	
-	
-}
