+++ src/cartas/CartaComAcao.java
@@ -1,0 +1,12 @@
+package cartas;
+
+public class CartaComAcao extends Carta {
+	
+	//Acao do tipo Acao que vai vir do Grupo Ação:  
+	//Muda desde Carta ou altera apenas aqui ?
+ 	
+	public int getNumero() {
+		//Dar um jeito de tratar essa retorno
+		return -1;
+	}
+}
