+++ src/cartas/CartaComCor.java
@@ -1,0 +1,12 @@
+package cartas;
+
+public class CartaComCor extends Carta {
+
+	
+	
+	public String getAcao() {
+		return "Esta carta não possui ação";
+	}
+	
+}
+
