+++ src/cartas/CartaEspecialComCor.java
@@ -1,0 +1,15 @@
+package cartas;
+
+public class CartaEspecialComCor extends CartaComAcao {
+	
+
+	public CartaEspecialComCor(String c, String ac){
+		
+		super.setCor(c);
+		super.setNumero(-1);
+		super.setAcao(ac);
+	}
+	
+	
+
+}
