+++ src/cartas/CartaNormal.java
@@ -1,0 +1,17 @@
+package cartas;
+
+public class CartaNormal extends CartaComCor {
+	
+	public CartaNormal(String c, int n){//Carta normal com cor
+		super.setCor(c);
+		super.setNumero(n);
+		super.setAcao(null);
+	}
+	
+	
+	
+	
+	
+	
+	
+}
